RF_Django/requirements.txt
coverage==5.5
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4